Rename weights functions
